Refactored the http method testing into tests kits to share code.
