+++ http-builder-ng-core/src/test/groovy/groovyx/net/http/UsesJavaClient.groovy
@@ -18,25 +18,11 @@ package groovyx.net.http
 import java.util.function.Function
 
 /**
- * Helpers for testing with the Apache client HttpBuilder.
+ * Denotes testing with the Java default client library.
  */
-class ApacheClientTesting {
+trait UsesJavaClient {
 
-    static final Function clientFactory = { c -> new ApacheHttpBuilder(c); } as Function
-
-    static HttpBuilder httpBuilder(Closure config) {
-        HttpBuilder.configure(clientFactory, config)
-    }
-
-    static HttpBuilder httpBuilder(String uri) {
-        httpBuilder {
-            request.uri = uri
-        }
-    }
-
-    static HttpBuilder httpBuilder(int port) {
-        httpBuilder {
-            request.uri = "http://localhost:${port}"
-        }
+    def setup() {
+        clientFactory = { c -> new JavaHttpBuilder(c) } as Function
     }
 }
